In this class, printAnagrams() is called, which prints all the words that can be formed using the letters of the input word. To do this, the number of occurrences of each letter in the input word is counted using a HashMap, and then all possible words are created and printed using the printAnagramsHelper() function.
The printAnagramsHelper() function constructs and prints all possible words using a HashMap and a character array. At each step, a letter is selected from the HashMap and the number of occurrences of that letter in the HashMap is reduced. Then this letter is placed in the character array and the function is called recursively to place other letters in the character array. At the end of each step, the number of occurrences of the selected letter in the HashMap is returned.